backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "MockMate backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.115.14",
    "sqlalchemy[asyncio]>=2.0.41",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "openai>=1.93.0",
]

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line-length = 100

[tool.mypy]
plugins = "sqlalchemy.ext.mypy.plugin"
